almost done with frontend
data is going and data is coming as expected 1. show the data on page 2. if for any case we fail to go through are lines of code in any stage then we will have a a repo file in are system so make make sure to  take care of it